+++ src/edu/bitsgoa/brokerfairness/util/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * this package contains some helper classes which are directly not part of project but may be used somewhere
+ */
+package edu.bitsgoa.brokerfairness.util;
+
+//rest of the file is empty